fully implemented order logic; created total_price view;

+++ src/main/kotlin/com/sorsix/barmanagmentapi/domain/views/OrderView.kt
@@ -29,7 +29,7 @@ data class OrderView(
     @Column(name = "drink_category")
     val drinkCategory: String,
     @Column(name = "drink_price")
-    val drinkPrice: Int,
+    val drinkPrice: Double,
     @Column(name = "quantity")
     val quantity: Int
 )
